accounts/abi: reorganizing package with small fixes (#14610)
* accounts/abi: reorganizing package and some notes and a quick correction of name.


get rid of some imports


* accounts/abi: move file names


* accounts/abi: fix boolean decode function


* accounts/abi: fix for the array set and for creating a bool


* accounts/abi: be very very very correct


* accounts/abi: fix up error message and variable names


* accounts/abi: take out unnecessary argument in pack method


* accounts/abi: add bool unpack test and add a panic to readBool function


* accounts/abi: fix panic message


* accounts/abi: change from panic to basic error


* accounts/abi: fix nil to false


* accounts/abi: fill out type regex tests and fill with the correct type for integers


* accounts/abi: move packNumbers into pack.go.


* accounts/abi: separation of the testing suite into appropriately named files.


* account/abi: change to hex string tests.


* account/abi: fix up rest of tests to hex


* accounts/abi: declare bool at the package level


* accounts/abi: use errors package in the error file.


* accounts/abi: fix ugly hack and fix error type declaration.
